requirements.txt
aerich==0.7.1
asyncpg==0.27.0
fastapi==0.94.1
ormar==0.12.1
passlib[bcrypt]
pre-commit==3.4.0
psycopg2-binary==2.9.5
tortoise-orm==0.19.3
uvicorn==0.21.1
